% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-funcs.R
\name{normalize_column_names}
\alias{normalize_column_names}
\title{normalize_column_names}
\usage{
normalize_column_names(data_)
}
\description{
changes the column names from TSV/CSV into legal tidyverse
colnames. e.g. replaces whitespaces with "." or removes characters that are
not valid in tibble colnames
}
