% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-funcs.R
\name{write_to_spreadsheet}
\alias{write_to_spreadsheet}
\title{write_to_spreadsheet}
\usage{
write_to_spreadsheet(data, file)
}
\description{
writes a list of named tibbles into XLSX via the openxlsx
package. Given names will become page labels.
}
\examples{
data = list(
   "a" = runif(10) \%>\% tibble::enframe(),
   "b" = runif(10) \%>\% tibble::enframe()
) \%>\% write.to.spreadsheet("test.xlsx")
}
